+++ internal/campaigns/features_test.go
@@ -2,6 +2,7 @@ package campaigns
 
 func featuresAllEnabled() featureFlags {
 	return featureFlags{
+		allowArrayEnvironments:   true,
 		includeAutoAuthorDetails: true,
 		useGzipCompression:       true,
 	}